Moving the useDataModelSchema to a more logical path
