+++ src/hooks/queries/useCurrentDataModelSchemaQuery.ts
@@ -4,8 +4,8 @@ import type { JSONSchema7 } from 'json-schema';
 
 import { useAppQueriesContext } from 'src/contexts/appQueriesContext';
 import { DataModelActions } from 'src/features/datamodel/datamodelSlice';
+import { useCurrentDataModelName } from 'src/features/datamodel/useDataModelSchema';
 import { useAppDispatch } from 'src/hooks/useAppDispatch';
-import { useCurrentDataModelName } from 'src/hooks/useDataModelSchema';
 import type { HttpClientError } from 'src/utils/network/sharedNetworking';
 
 export const useCurrentDataModelSchemaQuery = (): UseQueryResult<JSONSchema7> => {
