+++ src/hooks/useBindingSchema.tsx
@@ -1,6 +1,6 @@
 import type { JSONSchema7 } from 'json-schema';
 
-import { useDataModelSchema } from 'src/hooks/useDataModelSchema';
+import { useDataModelSchema } from 'src/features/datamodel/useDataModelSchema';
 import type { IDataModelBindings } from 'src/layout/layout';
 
 type AsSchema<T> = {
