+++ src/index.tsx
@@ -18,8 +18,8 @@ import { App } from 'src/App';
 import { ErrorBoundary } from 'src/components/ErrorBoundary';
 import { ThemeWrapper } from 'src/components/ThemeWrapper';
 import { AppQueriesContextProvider } from 'src/contexts/appQueriesContext';
+import { DataModelSchemaContextWrapper } from 'src/features/datamodel/useDataModelSchema';
 import { DevTools } from 'src/features/devtools/DevTools';
-import { DataModelSchemaContextWrapper } from 'src/hooks/useDataModelSchema';
 import * as queries from 'src/queries/queries';
 import { initSagas } from 'src/redux/sagas';
 import { setupStore } from 'src/redux/store';
