+++ src/testUtils.tsx
@@ -10,7 +10,7 @@ import type { PreloadedState } from 'redux';
 
 import { getInitialStateMock } from 'src/__mocks__/initialStateMock';
 import { AppQueriesContextProvider } from 'src/contexts/appQueriesContext';
-import { DataModelSchemaContextWrapper } from 'src/hooks/useDataModelSchema';
+import { DataModelSchemaContextWrapper } from 'src/features/datamodel/useDataModelSchema';
 import { setupStore } from 'src/redux/store';
 import { AltinnAppTheme } from 'src/theme/altinnAppTheme';
 import { ExprContextWrapper, useResolvedNode } from 'src/utils/layout/ExprContext';
